Fixed a deadlock between guiObjects and hotZones.
The mouse click handler was synchronizing on hotZones to iterate over it and then calling resetHideCouter which called scoreBoard.setShowMouseControls. setShowMouseControls synchronizes on guiObjects for iteration.

The paint method of scoreboard was locking guiObjects to iterate over it and then hotZones to stop a click event occuring during mid paint (when hotZones could be empty), it sounds unlikely but this was happening at high framerates.

The answer was of course to move scoreBoard.setShowMouseControls outside of the hotZones synchronization, just shows how easy it is to dead lock :(