Player and Game now raise events in new Threads.
Hopefully this will kill the deadlock issues once and for all. If not I could make each event have it's own thread or use CopyOnWriteArrayLists instead. I don't want to do the latter as it could lead to confusing things like an event being caught by an object after it's been removed from the listener list.